requirements.txt
requests~=2.28.1
beautifulsoup4~=4.11.1
pillow~=9.4.0
tqdm~=4.65.0